CTFd/plugins/ctfd_glowworm/vulns/web_pyblog/home/ctf/blog/requirements.txt
-i https://mirrors.tencent.com/pypi/simple
Django==1.11.22
Markdown==2.5.2
Pygments==2.2.0
psycopg2==2.7.1
MySQL-Python==1.2.5
PyYAML==3.13
